cpp bot actually plays now
